requirements_cuda8.txt
absl-py==0.7.1
albumentations==0.3.0
asn1crypto==0.24.0
astor==0.8.0
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
cloudpickle==1.2.1
colorama==0.4.1
#correlation-cuda==0.0.0
cryptography==2.7
cycler==0.10.0
dask==0.18.2
decorator==4.4.0
#Forward-Warp==0.0.1
#forward-warp-cuda==0.0.0
gast==0.2.2
grpcio==1.21.1
h5py==2.9.0
idna==2.8
imageio==2.5.0
imgaug==0.2.6
joblib==0.13.2
kiwisolver==1.1.0
libarchive-c==2.8
Markdown==3.0.1
matplotlib==2.2.4
mock==3.0.5
networkx==2.2
numpy==1.16.4
opencv-python==4.1.0.25
opencv-python-headless==4.1.0.25
Pillow==5.4.1
protobuf==3.8.0
pycparser==2.19
pyOpenSSL==19.0.0
pyparsing==2.4.0
pypng==0.0.15
PySocks==1.7.0
python-dateutil==2.8.0
pytz==2019.1
PyWavelets==1.0.3
PyYAML==5.1.1
requests==2.21.0
ruamel.yaml==0.15.46
scikit-image==0.14.3
scikit-learn==0.20.4
scipy==1.2.0
six==1.12.0
sklearn==0.0
#spatial-correlation-sampler==0.2.1
tensorboardX==1.0
tensorflow-estimator==1.13.0
termcolor==1.1.0
toolz==0.9.0
http://download.pytorch.org/whl/cu80/torch-0.4.1-cp36-cp36m-linux_x86_64.whl
#torch==0.4.1
torchvision==0.2.0
tqdm==4.32.1
urllib3==1.24.2
Werkzeug==0.15.4
